+++ libavfilter/vf_xpsnr.c
@@ -324,7 +324,7 @@ static inline double calcSquaredErrorAndWeight (AVFilterContext *ctx, XPSNRConte
   *msAct *= *msAct; /* because SSE is squared */ /* R&N: alpha k*/
   //TODO: R&N Delete begin
   gettimeofday(&end, 0);
-  av_log(ctx, AV_LOG_INFO, "******calcSquaredErrorAndWeight: differnt: %d ******\n", (long long)(end.tv_sec - begin.tv_sec)); 
+  av_log(ctx, AV_LOG_INFO, "******calcSquaredErrorAndWeight: differnt: %ld ******\n", (long)(end.tv_sec - begin.tv_sec)); 
 
   //TODO: R&N Delete end
   /* return nonweighted sum of squared errors */
@@ -491,7 +491,7 @@ static int getWSSE (AVFilterContext *ctx, int16_t **org, int16_t **orgM1, int16_
   } /* for c */
   //TODO: R&N Delete begin
   gettimeofday(&end, 0);
-  av_log(ctx, AV_LOG_INFO, "******getWSSE: differnt: %d ******\n", (long long)(end.tv_sec - begin.tv_sec)); 
+  av_log(ctx, AV_LOG_INFO, "******getWSSE: differnt: %ld ******\n", (long)(end.tv_sec - begin.tv_sec)); 
   //TODO: R&N Delete end
   return 0;
 }
@@ -606,7 +606,7 @@ static int do_xpsnr (FFFrameSync *fs)
     fprintf (s->statsFile, "\n");
   }  //TODO: R&N Delete begin 
   gettimeofday(&end, 0);
-  av_log(ctx, AV_LOG_INFO, "******do_xpsnr: differnt: %d ******\n", (long long)(end.tv_sec - begin.tv_sec));
+  av_log(ctx, AV_LOG_INFO, "******do_xpsnr: differnt: %ld ******\n", (long)(end.tv_sec - begin.tv_sec));
   //TODO: R&N Delete end  
   return ff_filter_frame (ctx->outputs[0], master);
 }
@@ -660,7 +660,7 @@ static av_cold int init (AVFilterContext *ctx)
 
   s->fs.on_event = do_xpsnr;  //TODO: R&N Delete begin
   gettimeofday(&end, 0);
-  av_log(ctx, AV_LOG_INFO, "******init: differnt: %d ******\n", (long long)(end.tv_sec - begin.tv_sec));
+  av_log(ctx, AV_LOG_INFO, "******init: differnt: %ld ******\n", (long)(end.tv_sec - begin.tv_sec));
   //TODO: R&N Delete end
   return 0;
 }
@@ -693,7 +693,7 @@ static int query_formats (AVFilterContext *ctx)
   if (fmts_list == NULL) return AVERROR (ENOMEM);
   //TODO: R&N Delete begin
   gettimeofday(&end, 0);
-  av_log(ctx, AV_LOG_INFO, "******query_formats: differnt: %d ******\n", (long long)(end.tv_sec - begin.tv_sec));
+  av_log(ctx, AV_LOG_INFO, "******query_formats: differnt: %ld ******\n", (long)(end.tv_sec - begin.tv_sec));
   //TODO: R&N Delete end
   return ff_set_common_formats (ctx, fmts_list);
 }
@@ -765,7 +765,7 @@ static int config_input_ref (AVFilterLink *inLink)
   }
   //TODO: R&N Delete begin
   gettimeofday(&end, 0);
-  av_log(inLink->dst, AV_LOG_INFO, "******config_input_ref: differnt: %d ******\n", (long long)(end.tv_sec - begin.tv_sec));
+  av_log(inLink->dst, AV_LOG_INFO, "******config_input_ref: differnt: %ld ******\n", (long)(end.tv_sec - begin.tv_sec));
   //TODO: R&N Delete end  
   return 0;
 }
@@ -793,7 +793,7 @@ static int config_output (AVFilterLink *outLink)
   if ((retValue = ff_framesync_configure (&s->fs)) < 0) return retValue;
   //TODO: R&N Delete begin
   gettimeofday(&end, 0);
-  av_log(outLink->src, AV_LOG_INFO, "******config_output: differnt: %d ******\n", (long long)(end.tv_sec - begin.tv_sec));
+  av_log(outLink->src, AV_LOG_INFO, "******config_output: differnt: %ld ******\n", (long)(end.tv_sec - begin.tv_sec));
   //TODO: R&N Delete end
   return 0;
 }
@@ -809,7 +809,7 @@ static int activate (AVFilterContext *ctx)
   //TODO: R&N Delete end
   return ff_framesync_activate (&s->fs);  //TODO: R&N Delete begin
   gettimeofday(&end, 0);
-  av_log(ctx, AV_LOG_INFO, "******activate: differnt: %d ******\n", (long long)(end.tv_sec - begin.tv_sec));
+  av_log(ctx, AV_LOG_INFO, "******activate: differnt: %ld ******\n", (long)(end.tv_sec - begin.tv_sec));
   //TODO: R&N Delete end
 }
 
@@ -887,7 +887,7 @@ static av_cold void uninit (AVFilterContext *ctx)
   //TODO: R&N Delete begin
   // print the difference between now and current time
   gettimeofday(&end, 0);
-  av_log(ctx, AV_LOG_INFO, "******uninit: differnt: %d ******\n", (long long)(end.tv_sec - begin.tv_sec));
+  av_log(ctx, AV_LOG_INFO, "******uninit: differnt: %ld ******\n", (long)(end.tv_sec - begin.tv_sec));
   //TODO: R&N Delete end
 }
